+++ src/driver/communication/usart/usart1_driver.h
@@ -33,13 +33,13 @@ void usart1_driver_power_off(void);
  * this function gives information if and how many bytes have been received
  * since the last time this function was called
  */
-u8 usart1_driver_bytes_available(void);
+u16 usart1_driver_bytes_available(void);
 
 /*!
  * get at maximum n bytes from the trx's modules internal buffer (the sw-buffer)
  * returns the number of bytes that have been copied into the given buffer
  */
-u8 usart1_driver_get_N_bytes(u8 num_bytes, u8* p_buffer_to);
+u16 usart1_driver_get_N_bytes(u16 num_bytes, u8* p_buffer_to);
 
 /*!
  * copys at maximum n bytes into the trx-modules' internal buffer
@@ -47,7 +47,7 @@ u8 usart1_driver_get_N_bytes(u8 num_bytes, u8* p_buffer_to);
  * this function does not start any operation, only the buffer is prepared.
  * returns the number of byts that have been copied into the given buffer
  */
-u8 usart1_driver_set_N_bytes(u8 num_bytes, const u8* p_buffer_from);
+u16 usart1_driver_set_N_bytes(u16 num_bytes, const u8* p_buffer_from);
 
 /*!
  *
@@ -67,7 +67,7 @@ void usart1_driver_start_rx(u16 num_of_rx_bytes);
  * @param num_bytes
  * @param timeout_ms
  */
-void usart1_driver_wait_for_rx(u8 num_bytes, u16 timeout_ms);
+void usart1_driver_wait_for_rx(u16 num_bytes, u16 timeout_ms);
 
 /*!
  * stops any active rx-operation
@@ -92,7 +92,7 @@ void usart1_driver_start_tx(void);
  * @param num_bytes
  * @param timeout_ms
  */
-void usart1_driver_wait_for_tx(u8 num_bytes, u16 timeout_ms);
+void usart1_driver_wait_for_tx(u16 num_bytes, u16 timeout_ms);
 
 /*!
  * stops any active tx-operation, internal buffer will left unchanged
