+++ src/system/system_interface.c
@@ -36,17 +36,17 @@ __UNUSED__ static void empty_driver_power_off(void) {
 
 }
 
-__UNUSED__ static u8 empty_driver_bytes_available (void) {
+__UNUSED__ static u16 empty_driver_bytes_available (void) {
 	return 0;
 }
 
-__UNUSED__ static u8 empty_driver_get_N_bytes (u8 num_bytes, u8* p_buffer_to) {
+__UNUSED__ static u16 empty_driver_get_N_bytes (u16 num_bytes, u8* p_buffer_to) {
 	(void) num_bytes;
 	(void) p_buffer_to;
 	return 0;
 }
 
-__UNUSED__ static u8 empty_driver_set_N_bytes (u8 num_bytes, const u8* p_buffer_from) {
+__UNUSED__ static u16 empty_driver_set_N_bytes (u16 num_bytes, const u8* p_buffer_from) {
 	(void) num_bytes;
 	(void) p_buffer_from;
 	return 0;
@@ -64,7 +64,7 @@ __UNUSED__ static void empty_driver_start_rx (u16 num_of_rx_bytes) {
 	(void) num_of_rx_bytes;
 }
 
-__UNUSED__ static void empty_driver_wait_for_rx(u8 num_bytes, u16 timeout_ms) {
+__UNUSED__ static void empty_driver_wait_for_rx(u16 num_bytes, u16 timeout_ms) {
 	(void) num_bytes;
 	(void) timeout_ms;
 }
@@ -75,7 +75,7 @@ __UNUSED__ static void empty_driver_stop_rx (void) {
 __UNUSED__ static void empty_driver_start_tx (void) {
 }
 
-__UNUSED__ static void empty_driver_wait_for_tx(u8 num_bytes, u16 timeout_ms) {
+__UNUSED__ static void empty_driver_wait_for_tx(u16 num_bytes, u16 timeout_ms) {
 	(void) num_bytes;
 	(void) timeout_ms;
 }
@@ -101,19 +101,19 @@ __UNUSED__ static void empty_driver_mutex_release(u8 m_id) {
 
 }
 
-__UNUSED__ void empty_driver_set_rx_bytes(u8 num_bytes, u8* p_buffer_from) {
+__UNUSED__ void empty_driver_set_rx_bytes(u16 num_bytes, u8* p_buffer_from) {
 	(void) num_bytes;
 	(void) p_buffer_from;
 
 }
 
-__UNUSED__ u8 empty_driver_get_tx_bytes(u8 num_bytes, u8* p_buffer_to) {
+__UNUSED__ u16 empty_driver_get_tx_bytes(u16 num_bytes, u8* p_buffer_to) {
 	(void) num_bytes;
 	(void) p_buffer_to;
 	return 0;
 }
 
-__UNUSED__ u8 empty_driver_get_tx_bytes_num_available(void) {
+__UNUSED__ u16 empty_driver_get_tx_bytes_num_available(void) {
 	return 0;
 }
